optimisation des patch avec turf
