+++ package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/ign-packo/PackO#readme",
   "dependencies": {
+    "@turf/turf": "^6.3.0",
     "body-parser": "^1.19.0",
     "c8": "^7.3.5",
     "chai": "^4.2.0",
